Ajuste de estructura para week1day2
Se agregan, mueven y eliminan carpetas para poder ajustar la estructura del proyecto a la sugerencia del profesor; por ejemplo, la carpeta modules y sus archivos se eliminaron, la carpeta de excepciones se agregó y se movió la carpeta de interfaces a persistance/repositories
